Handle already memoized strings and BINGET
Under certain conditions a string may have already been "memoized"
during the pickling process.  If this is the case, the value during
opcode sequence unwinding is handled by reference (the offset within
the "memo") rather than by value.

The implementation here is a shortcut as we do not care about
decoding the entire Django session but rather just the OMERO.web
Connector.

See:
 * https://github.com/python/cpython/blob/3.10/Lib/pickletools.py