OmeroRequest: close resources on client.joinSession() failure
As described in https://docs.oracle.com/javase/specs/jls/se8/html/jls-14.html#jls-14.20.3,
a resource is closed only if it initialized to a non-null value.
In the rare event of the client.joinSession() call throwing an exception,
the OmeroRequest.close() statement will never be invoked and clean all
resources associated with the session joining operations.
Given the semantics of the omero.client object, this can cause an
accumulation of objects within a micro-service process which would only
be cleaned up by a restart.

This commit handles this scenario by catching the exception in the constructor,
closing the resources and throwing a CannotCreateSessionException exception which
is already part of the signature.